+++ src/WebDriver.BiDi/Modules/BrowsingContext/TraverseHistoryCommand.cs
@@ -4,7 +4,7 @@ namespace OpenQA.Selenium.BiDi.Modules.BrowsingContext;
 
 internal class TraverseHistoryCommand(TraverseHistoryCommandParameters @params) : Command<TraverseHistoryCommandParameters>(@params);
 
-internal record TraverseHistoryCommandParameters(BrowsingContext Context, int Delta) : CommandParameters;
+internal record TraverseHistoryCommandParameters(BrowsingContext Context, long Delta) : CommandParameters;
 
 public record TraverseHistoryOptions : CommandOptions;
