+++ src/WebDriver.BiDi/Modules/Network/AuthRequiredEventArgs.cs
@@ -2,5 +2,5 @@
 
 namespace OpenQA.Selenium.BiDi.Modules.Network;
 
-public record AuthRequiredEventArgs(BiDi.Session Session, BrowsingContext.BrowsingContext Context, bool IsBlocked, BrowsingContext.Navigation Navigation, uint RedirectCount, RequestData Request, DateTimeOffset Timestamp, ResponseData Response) :
+public record AuthRequiredEventArgs(BiDi.Session Session, BrowsingContext.BrowsingContext Context, bool IsBlocked, BrowsingContext.Navigation Navigation, ulong RedirectCount, RequestData Request, DateTimeOffset Timestamp, ResponseData Response) :
     BaseParametersEventArgs(Session, Context, IsBlocked, Navigation, RedirectCount, Request, Timestamp);
