+++ src/WebDriver.BiDi/Modules/Network/FetchErrorEventArgs.cs
@@ -3,5 +3,5 @@ using System;
 
 namespace OpenQA.Selenium.BiDi.Modules.Network;
 
-public record FetchErrorEventArgs(BiDi.Session Session, BrowsingContext.BrowsingContext Context, bool IsBlocked, Navigation Navigation, uint RedirectCount, RequestData Request, DateTimeOffset Timestamp, string ErrorText)
+public record FetchErrorEventArgs(BiDi.Session Session, BrowsingContext.BrowsingContext Context, bool IsBlocked, Navigation Navigation, ulong RedirectCount, RequestData Request, DateTimeOffset Timestamp, string ErrorText)
     : BaseParametersEventArgs(Session, Context, IsBlocked, Navigation, RedirectCount, Request, Timestamp);
