+++ src/WebDriver.BiDi/Modules/Network/RequestData.cs
@@ -2,4 +2,4 @@
 
 namespace OpenQA.Selenium.BiDi.Modules.Network;
 
-public record RequestData(Request Request, string Url, string Method, IReadOnlyList<Header> Headers, IReadOnlyList<Cookie> Cookies, uint HeadersSize, uint? BodySize, FetchTimingInfo Timings);
+public record RequestData(Request Request, string Url, string Method, IReadOnlyList<Header> Headers, IReadOnlyList<Cookie> Cookies, ulong HeadersSize, ulong? BodySize, FetchTimingInfo Timings);
